Extract live send run composition factories
