+++ src/PlateauResoniteLink/Targets/Resonite/ResoniteLiveSceneImportDependencyFactory.cs
@@ -1,0 +1,39 @@
+using System;
+using System.Net.Http;
+
+using PlateauResoniteLink.Transport.ResoniteLink;
+
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal interface IResoniteLiveSceneImportDependencyFactory
+{
+    ResoniteLiveSceneImportDependencies Create(
+        ResoniteLiveSceneImportTargetOptions options,
+        HttpClient terrainTextureAssetHttpClient);
+}
+
+internal sealed class ResoniteLiveSceneImportDependencyFactory(
+    IResoniteClientSessionFactory clientSessionFactory,
+    IResoniteLiveSendRunStarterFactory runStarterFactory,
+    IResoniteLiveSendStartRequestFactory startRequestFactory,
+    IResoniteLiveSendQueue queue)
+    : IResoniteLiveSceneImportDependencyFactory
+{
+    public ResoniteLiveSceneImportDependencies Create(
+        ResoniteLiveSceneImportTargetOptions options,
+        HttpClient terrainTextureAssetHttpClient)
+    {
+        ArgumentNullException.ThrowIfNull(options);
+        ArgumentNullException.ThrowIfNull(terrainTextureAssetHttpClient);
+        ResoniteLinkSendDiagnostics diagnostics = options.EnableSendMetrics
+            ? ResoniteLinkSendDiagnostics.CreateEnabled(options.ProgressReporter)
+            : ResoniteLinkSendDiagnostics.Disabled;
+
+        return new ResoniteLiveSceneImportDependencies(
+            clientSessionFactory.Create(options, diagnostics),
+            diagnostics,
+            startRequestFactory,
+            runStarterFactory.Create(terrainTextureAssetHttpClient, options),
+            queue);
+    }
+}
