+++ src/PlateauResoniteLink/Targets/Resonite/ResoniteLiveSceneImportFactory.cs
@@ -1,0 +1,24 @@
+using System.Net.Http;
+
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal interface IResoniteLiveSceneImportFactory
+{
+    ResoniteLiveSceneImportTarget CreateTarget(
+        ResoniteLiveSceneImportTargetOptions options,
+        HttpClient terrainTextureAssetHttpClient);
+}
+
+internal sealed class ResoniteLiveSceneImportFactory(
+    IResoniteLiveSceneImportDependencyFactory dependencyFactory) : IResoniteLiveSceneImportFactory
+{
+    public ResoniteLiveSceneImportTarget CreateTarget(
+        ResoniteLiveSceneImportTargetOptions options,
+        HttpClient terrainTextureAssetHttpClient)
+    {
+        ResoniteLiveSceneImportDependencies dependencies = dependencyFactory.Create(
+            options,
+            terrainTextureAssetHttpClient);
+        return new ResoniteLiveSceneImportTarget(options, dependencies);
+    }
+}
