+++ src/PlateauResoniteLink/Targets/Resonite/ResoniteLiveSendTargetServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Net.Http;
 
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
@@ -26,8 +25,11 @@ public static class ResoniteLiveSendTargetServiceCollectionExtensions
         services.TryAddScoped<IResoniteGeometryAssetPlanner, ResoniteGeometryAssetPlanner>();
         services.TryAddScoped<IResoniteMaterialPlanning, ResoniteMaterialPlanning>();
         services.TryAddScoped<IResoniteSceneMaterialPlanComposer, ResoniteSceneMaterialPlanComposer>();
+        services.TryAddScoped<IResoniteCommonMaterialSetupPreparer, ResoniteCommonMaterialSetupPreparer>();
         services.TryAddScoped<ILiveSendRunPlanFactory, LiveSendRunPlanFactory>();
         services.TryAddScoped<ILiveSendRunStateFactory, LiveSendRunStateFactory>();
+        services.TryAddScoped<IResoniteLiveSendRunStarterFactory, ResoniteLiveSendRunStarterFactory>();
+        services.TryAddScoped<IResoniteLiveSendWorkerLauncherFactory, ResoniteLiveSendWorkerLauncherFactory>();
         services.TryAddScoped<IResoniteLiveSendStartRequestFactory, ResoniteLiveSendStartRequestFactory>();
         services.TryAddScoped<IResonitePreparedCityObjectImporter, ResonitePreparedCityObjectImporter>();
         services.TryAddScoped<IResoniteQueuedCityObjectEnqueuer, ResoniteQueuedCityObjectEnqueuer>();
@@ -48,128 +50,5 @@ public static class ResoniteLiveSendTargetServiceCollectionExtensions
         services.TryAddScoped<IResoniteLiveSceneImportFactory, ResoniteLiveSceneImportFactory>();
 
         return services;
-    }
-}
-
-internal interface IResoniteLiveSceneImportFactory
-{
-    ResoniteLiveSceneImportTarget CreateTarget(
-        ResoniteLiveSceneImportTargetOptions options,
-        HttpClient terrainTextureAssetHttpClient);
-}
-
-internal sealed class ResoniteLiveSceneImportFactory(
-    IResoniteLiveSceneImportDependencyFactory dependencyFactory) : IResoniteLiveSceneImportFactory
-{
-    public ResoniteLiveSceneImportTarget CreateTarget(
-        ResoniteLiveSceneImportTargetOptions options,
-        HttpClient terrainTextureAssetHttpClient)
-    {
-        ResoniteLiveSceneImportDependencies dependencies = dependencyFactory.Create(
-            options,
-            terrainTextureAssetHttpClient);
-        return new ResoniteLiveSceneImportTarget(options, dependencies);
-    }
-}
-
-internal interface IResoniteLiveSceneImportDependencyFactory
-{
-    ResoniteLiveSceneImportDependencies Create(
-        ResoniteLiveSceneImportTargetOptions options,
-        HttpClient terrainTextureAssetHttpClient);
-}
-
-internal interface IResoniteClientSessionFactory
-{
-    ILiveSendClientSession Create(
-        ResoniteLiveSceneImportTargetOptions options,
-        ResoniteLinkSendDiagnostics diagnostics);
-}
-
-internal sealed class ResoniteLinkClientSessionFactory(
-    Func<Action<string>?, IResoniteLinkClient> baseClientFactory) : IResoniteClientSessionFactory
-{
-    public ILiveSendClientSession Create(
-        ResoniteLiveSceneImportTargetOptions options,
-        ResoniteLinkSendDiagnostics diagnostics)
-    {
-        ArgumentNullException.ThrowIfNull(options);
-        ArgumentNullException.ThrowIfNull(diagnostics);
-
-        return ResoniteLinkTransportSessionFactory.Create(
-            options.Endpoint,
-            options.ConnectionCount,
-            diagnostics,
-            options.ProgressReporter,
-            baseClientFactory);
-    }
-}
-
-internal interface ITerrainTextureAssetGeneratorFactory
-{
-    ITerrainTextureAssetGenerator Create(
-        HttpClient terrainTextureAssetHttpClient,
-        ResoniteLiveSceneImportTargetOptions options);
-}
-
-internal sealed class TerrainTextureAssetGeneratorFactory : ITerrainTextureAssetGeneratorFactory
-{
-    public ITerrainTextureAssetGenerator Create(
-        HttpClient terrainTextureAssetHttpClient,
-        ResoniteLiveSceneImportTargetOptions options)
-    {
-        ArgumentNullException.ThrowIfNull(terrainTextureAssetHttpClient);
-        ArgumentNullException.ThrowIfNull(options);
-        return new TerrainTextureAssetGenerator(
-            terrainTextureAssetHttpClient,
-            options.TerrainTileCacheRoot,
-            options.DisableTerrainTileCache);
-    }
-}
-
-internal sealed class ResoniteLiveSceneImportDependencyFactory(
-    IResoniteClientSessionFactory clientSessionFactory,
-    ITerrainTextureAssetGeneratorFactory terrainTextureAssetGeneratorFactory,
-    IResoniteSceneSetupInterpreter sceneSetupInterpreter,
-    IResoniteDatasetLicenseWriter datasetLicenseWriter,
-    IResoniteMaterialPlanning materialPlanning,
-    ILiveSendRunPlanFactory runPlanFactory,
-    ILiveSendRunStateFactory runStateFactory,
-    IResoniteLiveSendStartRequestFactory startRequestFactory,
-    IResonitePreparedCityObjectImporter preparedCityObjectImporter,
-    IResoniteSlotCreator slotCreator,
-    IResoniteLiveSendQueue queue)
-    : IResoniteLiveSceneImportDependencyFactory
-{
-    public ResoniteLiveSceneImportDependencies Create(
-        ResoniteLiveSceneImportTargetOptions options,
-        HttpClient terrainTextureAssetHttpClient)
-    {
-        ArgumentNullException.ThrowIfNull(options);
-        ArgumentNullException.ThrowIfNull(terrainTextureAssetHttpClient);
-        ResoniteLinkSendDiagnostics diagnostics = options.EnableSendMetrics
-            ? ResoniteLinkSendDiagnostics.CreateEnabled(options.ProgressReporter)
-            : ResoniteLinkSendDiagnostics.Disabled;
-
-        ResoniteQueuedCityObjectSender queuedCityObjectSender = new(
-            terrainTextureAssetGeneratorFactory.Create(terrainTextureAssetHttpClient, options),
-            datasetLicenseWriter,
-            preparedCityObjectImporter);
-        ResoniteQueuedCityObjectWorker queuedCityObjectWorker = new(queuedCityObjectSender);
-        ResoniteLiveSendWorkerLauncher workerLauncher = new(queuedCityObjectWorker);
-        ResoniteLiveSendRunStarter runStarter = new(
-            sceneSetupInterpreter,
-            new ResoniteCommonMaterialSetupPreparer(materialPlanning, options.ProgressReporter),
-            runPlanFactory,
-            runStateFactory,
-            workerLauncher,
-            slotCreator);
-
-        return new ResoniteLiveSceneImportDependencies(
-            clientSessionFactory.Create(options, diagnostics),
-            diagnostics,
-            startRequestFactory,
-            runStarter,
-            queue);
     }
 }
